requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pytz==2020.4
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.2.0
